Refactor: move transformer functions to transformers.go
