+++ planner/utils.go
@@ -2,10 +2,7 @@ package planner
 
 import (
 	"errors"
-	"github.com/weihesdlegend/Vacation-planner/POI"
 	"regexp"
-	"strconv"
-	"time"
 )
 
 // validate date is in the format of yyyy-mm-dd
@@ -19,21 +16,6 @@ func validateDate(date string) error {
 		return errors.New("date format must be yyyy-mm-dd")
 	}
 	return nil
-}
-
-func toWeekday(date string) POI.Weekday {
-	datePattern := regexp.MustCompile(`(?P<year>\d{4})-(?P<month>\d{2})-(?P<day>\d{2})`)
-	dateFields := datePattern.FindStringSubmatch(date)
-	year, _ := strconv.Atoi(dateFields[1])
-	month, _ := strconv.Atoi(dateFields[2])
-	day, _ := strconv.Atoi(dateFields[3])
-	t := time.Date(year, time.Month(month), day, 0, 0, 0, 0, time.UTC)
-	return POI.Weekday(t.Weekday())
-}
-
-func toPriceLevel(priceLevel string) POI.PriceLevel {
-	price, _ := strconv.Atoi(priceLevel)
-	return POI.PriceLevel(price)
 }
 
 // validate location is in the format of city,country
